services/executor-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "growthpilot-executor-service"
version = "0.1.0"
description = "Consumes actions.proposed, applies guardrails, executes actions, emits actions.applied."
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6.1",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "aiokafka>=0.10.0",
  "orjson>=3.10.0",
  "python-dotenv>=1.0.1",
  "datadog>=0.49.1",
  "psycopg[binary]>=3.2.1",
  "psycopg-pool>=3.2.0",
  "httpx>=0.27.0",

  "growthpilot-contracts @ file:///contracts/python",
  "growthpilot-shared @ file:///shared/python",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]